+++ javaScript/projects/weather-app/src/components/tabs/forecast/forecast.config.js
@@ -23,38 +23,12 @@ const forecast = {
 };
 
 const forecastController = {
-  init(unitSystem, weatherData) {
-    this.weatherData = weatherData;
-    this.unitSystem = unitSystem;
-    this.findObjects = this.findObjects.bind(this);
-    Object.keys(forecast).forEach(this.findObjects);
-  },
-  setValues(key, subKey, item) {
-    if (item) {
-      forecast[key].push(item);
-    } else {
-      forecast[key][subKey] = this.weatherData[key][subKey];
-    }
-  },
-  findObjects(key) {
-    if (!(forecast[key] instanceof Array)) {
-      Object.keys(forecast[key]).forEach((subKey) => {
-        this.setValues(key, subKey);
-      });
-    } else {
-      this.weatherData.forecast[key].forEach((item) => {
-        this.setValues(key, null, item);
-      });
-    }
-  },
+  init(unitSystem, weatherData) {},
+  setValues(key, subKey, item) {},
+  findObjects(key) {},
 };
 
 export default {
-  init(unitSystem, weatherData) {
-    forecastController.init(unitSystem, weatherData);
-    this.setProperties();
-  },
-  setProperties() {
-    Object.assign(this, { ...forecast }, { location: forecastController.weatherData.location });
-  },
+  init(unitSystem, weatherData) {},
+  setProperties() {},
 };